Split out tests according to mutation
