+++ app/slack_integration/types.py
@@ -124,3 +124,12 @@ class SolveQuestionFromSlackInputType(graphene.InputObjectType):
     slack_channel_id = graphene.String()
     slack_user_id = graphene.String()
     time_end = graphene.String()
+
+
+class UserAndMessageFromSlackInputType(graphene.InputObjectType):
+    slack_user = graphene.Field(UserFromSlackInputType)
+    origin_slack_event_ts = graphene.String(required=True)
+    slack_channel_id = graphene.String(required=True)
+    slack_user_id = graphene.String(required=True)
+    text = graphene.String(required=True)
+    time = graphene.String(required=True)
