+++ tests/func/groups/test_query_groups.py
